mlops/airflow/requirements.txt
requests
minio
pendulum==2.1.2
python-magic
faster-whisper
ffmpeg
transformers
openai-clip
opencv-python
summa
argostranslate
opencv-python
pillow
